requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
plotly
scikit-learn
statsmodels
xgboost
sklearn-genetic-opt
joblib
scipy
deap
threadpoolctl
requests